refactor logic add tests
